+++ tasks/peripheralsTasks.ts
